+++ src/TodoContainer.module.css
@@ -165,7 +165,7 @@ body button:hover {
 }
 
 .footer-container {
-    height: 13vh;
+    height: 10vh;
     background-color: #aee6e285;
     margin-top: 27em;
 }
